Support F2003 array syntax and F2008 coarrays
Import from Geany: https://github.com/geany/geany/commit/1c0fa99e1d84b0331d65e56dcab4002f0006ccb1
Contributed by aphirst on 15 Feb 2014.

Fortran 2003 allows '[' and ']' as array constructors in addition to
the legacy '(/' and '/)'. Fortran 2008 allows '[]' as part of a
CODIMENSION specifier, which can either be `codimension[]` in
the type-spec; or as either `variable[]` or `variable()[]` in
the entry-spec.

Credit to Colomban Wendling for the skipOverSquares logic, and
treating `codimension` as a special-case.